Move different inde tracker tests to IndexTrackerTest.
